address review: single-cell content-control fix, move + expand tests
- Fix the 1x1 layout-table shortcut in _handle_tables: it used
  table.rows[0].cells[0], which walks only direct w:tc children, so a
  lone cell wrapped in a w:sdt content control was missed, cells[0]
  raised IndexError, and the caller silently dropped the cell content
  (the exact shape Word emits for cover-page / document-property
  controls). Route it through _row_cells() like the multi-cell path.
- Move the content-control test and its helper into
  test_backend_msword.py (drop the separate module) and add a
  type annotation to the helper's `table` parameter.
- Add a middle-column-only case (build((2,))) so the grid-shift is
  verified independently of the first cell being wrapped, and a
  regression test for the single-cell wrapped case.
