+++ tests/test_backend_msword.py
@@ -18,6 +18,7 @@ from docx.enum.style import WD_STYLE_TYPE
 from docx.opc.constants import RELATIONSHIP_TYPE as RT
 from docx.oxml import OxmlElement
 from docx.oxml.ns import qn
+from docx.table import Table
 from lxml import etree
 from PIL import Image
 
@@ -1494,3 +1495,94 @@ def test_docx_code_detection(tmp_path):
     assert not _code_items(doc), "A code-styled checkbox must not become a CodeItem"
     design = next(it for it in doc.texts if it.text == "Design")
     assert str(getattr(design.label, "value", design.label)).startswith("checkbox")
+
+
+def _wrap_cell_in_content_control(table: Table, row: int, col: int) -> None:
+    """Move a cell under ``w:sdt``/``w:sdtContent``, the way Word does.
+
+    Word produces this shape for date pickers and for cells bound to document
+    properties: the ``w:tc`` is no longer a direct child of the ``w:tr``.
+    """
+    tc = table.cell(row, col)._tc
+    tr = tc.getparent()
+    position = list(tr).index(tc)
+    sdt = tr.makeelement(qn("w:sdt"), {})
+    sdt_content = tr.makeelement(qn("w:sdtContent"), {})
+    tr.remove(tc)
+    sdt_content.append(tc)
+    sdt.append(sdt_content)
+    tr.insert(position, sdt)
+
+
+def test_table_cells_inside_a_content_control_keep_their_grid_column(tmp_path):
+    """A content-control cell must be parsed, and must not shift the row left.
+
+    Only direct ``w:tc`` children of a ``w:tr`` used to be visited, so a cell
+    Word had wrapped in a content control was dropped -- and because the grid
+    column advances once per emitted cell, every later cell in the row moved
+    into the vacated column and lined up under the wrong header.
+    """
+
+    converter = DocumentConverter(allowed_formats=[InputFormat.DOCX])
+    header = ["Date", "Version", "Author", "Note"]
+    row = ["1.2.2015", "1", "Acme s.r.o.", "Created"]
+
+    def build(wrapped_columns: tuple[int, ...]) -> list[str]:
+        doc = Document()
+        table = doc.add_table(rows=2, cols=len(header))
+        for col, value in enumerate(header):
+            table.cell(0, col).text = value
+        for col, value in enumerate(row):
+            table.cell(1, col).text = value
+        for col in wrapped_columns:
+            _wrap_cell_in_content_control(table, 1, col)
+
+        path = tmp_path / f"content_control_{'_'.join(map(str, wrapped_columns))}.docx"
+        doc.save(str(path))
+
+        data = converter.convert(path).document.tables[0].data
+        grid = [""] * data.num_cols
+        for cell in data.table_cells or []:
+            if cell.start_row_offset_idx == 1:
+                grid[cell.start_col_offset_idx] = cell.text or ""
+        return grid
+
+    assert build(()) == row
+    # A single wrapped cell in the first column: previously dropped, shifting
+    # the rest one column left.
+    assert build((0,)) == row
+    # Only a middle column wrapped: the grid-shift must be caught independently
+    # of the first cell being wrapped.
+    assert build((2,)) == row
+    # Two wrapped cells in the same row, one of them not the first.
+    assert build((0, 2)) == row
+
+
+def test_single_cell_layout_table_wrapped_in_content_control(tmp_path):
+    """A one-cell layout table whose only cell is content-control wrapped.
+
+    A ``1x1`` table is treated as furniture and its cell is walked as body
+    content. That shortcut used ``table.rows[0].cells[0]``, which walks only
+    direct ``w:tc`` children, so when the lone cell is wrapped in a ``w:sdt``
+    (the shape Word emits for cover-page and document-property controls) the
+    row had no cells, ``cells[0]`` raised ``IndexError``, and the caller's
+    ``except`` silently dropped the cell's content.
+    """
+
+    converter = DocumentConverter(allowed_formats=[InputFormat.DOCX])
+    doc = Document()
+    doc.add_paragraph("before")
+    table = doc.add_table(rows=1, cols=1)
+    table.cell(0, 0).text = "Cover value"
+    _wrap_cell_in_content_control(table, 0, 0)
+    doc.add_paragraph("after")
+
+    path = tmp_path / "single_cell_content_control.docx"
+    doc.save(str(path))
+
+    texts = [
+        item.text
+        for item, _ in converter.convert(path).document.iterate_items()
+        if isinstance(item, TextItem)
+    ]
+    assert "Cover value" in texts
